+++ platformio.ini
@@ -26,7 +26,7 @@ debug_tool = esp-prog
 debug_init_break = tbreak setup
 build_flags = 
     -D FASTLED_INTERNAL ; disable FastLED warning
-	-D CORE_DEBUG_LEVEL=4
+	-D CORE_DEBUG_LEVEL=5
     -D CONFIG_ARDUHAL_LOG_COLORS
 board_build.partitions = min_spiffs.csv
